Re-enable hd warning dialog and clean up settings
